File (re-written after lost data)
=== Main Notes ===
Rewritten the files and methods that I lost when I pulled from origin and overwrite earlier on.

=== Compile and Testing State ===
File.java is untested and not a point of compiling as I need to sort out dependancies

=== Changes ===
File -> game specific methods have some dependancies that I need to go through and sort out.

FixedTile.java is a extension/child of FloorTile with x and y positions, which I need for reading and writing the board.

Added afew getter methods in FloorTile, and Game.

Co-Authored-By: OJH99 <73654778+OJH99@users.noreply.github.com>